+++ src/builtin_plugins/amulet_team_3d_viewer/_view_3d/__init__.py
@@ -1,1 +1,1 @@
-from ._view import View3D
+from ._widget import View3D
