+++ src/builtin_plugins/amulet_team_3d_viewer/_view_3d/_widget.py
@@ -7,7 +7,7 @@ from PySide6.QtWidgets import QWidget, QVBoxLayout
 
 from amulet_team_main_window import TabWidget
 
-from ._renderer import FirstPersonCanvas
+from ._canvas import FirstPersonCanvas
 
 
 class View3D(TabWidget):
